Merger: Combine spec helpers
