+++ spec/spec_helper.rb
@@ -1,15 +1,33 @@
 # frozen_string_literal: true
 
+require "simplecov"
 require "coradoc"
 
 RSpec.configure do |config|
   # Enable flags like --only-failures and --next-failure
   config.example_status_persistence_file_path = ".rspec_status"
 
-  # Disable RSpec exposing methods globally on `Module` and `main`
-  config.disable_monkey_patching!
-
-  config.expect_with :rspec do |c|
-    c.syntax = :expect
+  SimpleCov.profiles.define "gem" do
+    add_filter "/spec/"
+    add_filter "/autotest/"
+    add_group "Libraries", "/lib/"
   end
+  SimpleCov.start "gem"
+
+  # ReverseAdoc:
+  config.after(:each) do
+    Coradoc::ReverseAdoc.instance_variable_set(:@config, nil)
+  end
+end
+
+# ReverseAdoc:
+require "coradoc/reverse_adoc"
+require "coradoc/reverse_adoc/html_converter"
+require "word-to-markdown"
+
+Dir[File.join("spec", "reverse_adoc", "support", "**", "*.rb")]
+  .each { |f| require File.join(".", f) }
+
+def node_for(html)
+  Nokogiri::HTML.parse(html).root.child.child
 end
